cicd_failure_prediction/backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
PyYAML>=6.0.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
python-dotenv>=1.0.0
werkzeug==3.0.1
